Bugfix: Closing Wireshark causes stacktrace/crash of internalblue.
The problem is, that the context.arch variable defaults to 'i386'
for every new python thread or timer. The timer which handles the
shutdown of Wireshark tries to remove the patches and therefore
calls asm() which will try to assemble intel code and fail. The
workaround/fix is to set the arch manually everytime we call asm.
By this, also external scripts don't need to set context.arch='thumb'
in every thread they start.